Make it work on Java 11
The error was that the XMLSignature.sign invocation was given the
original document, and mutated the document that was signed. For the
signing to work properly it is now given a _new_ empty Document. This
is a much cleaner approach, as the original document instance (which is signed)
is not mutated.

The overly awkward use of misunderstood object-orient'ness in these APIs of the JDK are confusing to say the
least. And don't get me started on the thread safety, or lack thereof, of all the factories and stuff.

Co-authored-by: Øyvind Nerbråten <oyvind@nerbraten.no>